Changed sorting
I changed the way the script sorts the photos. Instead of sorting by the capture time, it now sorts based on the original file's sequence number. This ensures that if shooting at a high FPS and the capture times are all the same, the script doesn't overwrite previous shots as it things it's the same shot. A potential risk would be if there are multiple shots from the same day from two cameras of the same make (Canon, Fujifilm, Sony, etc.) that also save a photo with the same photo sequence number.